Alerts Controller
index function: update status filter to be an array like the other filters in-use through the site. Update filter queries and data prep for the array being passed to the view.  AlertData now always returned as JSON. Vue component will pull the records in mounted() method... AFTER filters are loaded from the store.